WAS ADDED THE CLASS: - DAO CHANGED MAIN CLASSE

+++ api.capsrnrb/src/main/java/br/com/api/capsrnrb/models/Usuarios.java
@@ -23,7 +23,7 @@ public class Usuarios
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	@Column
 	private String nome;
@@ -50,13 +50,11 @@ public class Usuarios
 	private String nuTelefone;
 	
 		
-	public Integer getId()
-	{
+	public Long getId(){
 		return this.id;
 	}
 
-	public void setId(Integer id)
-	{
+	public void setId(Long id){
 		this.id = id;
 	}
